cloud_functions/news-api-ingest/requirements.txt
requests
google-cloud-bigquery
google-cloud